requirements.txt
fastapi
uvicorn
azure-servicebus
python-dotenv
